Split extractor plugin registry for #3422
