+++ changelog.d/unreleased/3422.internal.md
@@ -1,0 +1,19 @@
+---
+category: internal
+issues:
+  - 3422
+affected:
+  - src/CodeIndex/Indexer/Extensibility/ExtractorPluginRegistry.cs
+  - src/CodeIndex/Indexer/Extensibility/ExtractorPluginRegistry.Diagnostics.cs
+  - src/CodeIndex/Indexer/Extensibility/ExtractorPluginRegistry.Discovery.cs
+  - src/CodeIndex/Indexer/Extensibility/ExtractorPluginRegistry.PatternConfig.cs
+  - src/CodeIndex/Indexer/Extensibility/ExtractorPluginRegistry.PluginLoading.cs
+---
+
+## English
+
+- **Split extractor plugin registry responsibilities (#3422)** — `ExtractorPluginRegistry` now keeps discovery, pattern config handling, plugin assembly loading, and diagnostic reporting in focused partial files while preserving the existing extension behavior.
+
+## 日本語
+
+- **extractor plugin registry の責務を分割しました (#3422)** — `ExtractorPluginRegistry` は既存の extension 挙動を維持したまま、探索、pattern config 処理、plugin assembly loading、diagnostic reporting を責務別の partial file に分けました。
